# src/main/resources/openapi/catalogue.yaml
openapi: 3.0.3
info:
  title: School Management API
  version: 1.0.0
  description: API for managing catalogues, students, disciplines, and grades.

servers:
  - url: http://localhost:8081

tags:
  - name: Students
    description: Operations related to students
  - name: Admin
    description: Administrative operations
  - name: Professors
    description: Operations related to professors and grades

paths:
  /api/admin/students:
    post:
      tags:
        - Admin
      summary: Create a new student
      operationId: createStudent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStudentRequest'
      responses:
        '201':
          description: Student created
    get:
      tags:
        - Admin
      summary: Get all students
      operationId: getAllStudents
      responses:
        '200':
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentResponse'

  /api/admin/students/{id}:
    get:
      tags:
        - Admin
      summary: Get student by ID
      operationId: getStudentById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Student details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentResponse'
    put:
      tags:
        - Admin
      summary: Update student by ID
      operationId: updateStudent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStudentRequest'
      responses:
        '200':
          description: Updated student
    delete:
      tags:
        - Admin
      summary: Delete student by ID
      operationId: deleteStudent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Student deleted

  /api/students/classCode/{classCode}/professors:
    get:
      tags:
        - Students
      summary: Get professors for a class and year
      operationId: getProfessorsForStudent
      parameters:
        - in: path
          name: classCode
          required: true
          schema:
            type: string
        - in: query
          name: year
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Professors list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProfessorResponse'

  /api/students/studentCode/{studentCode}/grades:
    get:
      tags:
        - Students
      summary: Get grades by studentCode
      operationId: getGradesForStudent
      parameters:
        - in: path
          name: studentCode
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Grades list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GradesStudentResponse'

  /api/admin/disciplines:
    post:
      tags:
        - Admin
      summary: Create a new discipline
      operationId: createDiscipline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDisciplineRequest'
      responses:
        '201':
          description: Discipline created
    get:
      tags:
        - Admin
      summary: Get all disciplines
      operationId: getAllDisciplines
      responses:
        '200':
          description: List of disciplines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DisciplineResponse'

  /api/admin/disciplines/{id}:
    get:
      tags:
        - Admin
      summary: Get discipline by ID
      operationId: getDisciplineById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Discipline details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisciplineResponse'
    put:
      tags:
        - Admin
      summary: Update discipline by ID
      operationId: updateDiscipline
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDisciplineRequest'
      responses:
        '200':
          description: Updated discipline
    delete:
      tags:
        - Admin
      summary: Delete discipline by ID
      operationId: deleteDiscipline
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Discipline deleted

  /api/professors/grades:
    post:
      tags:
        - Professors
      summary: Create a new grade
      operationId: createGrade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGradeRequest'
      responses:
        '201':
          description: Grade created
    get:
      tags:
        - Professors
      summary: Get all grades
      operationId: getAllGrades
      responses:
        '200':
          description: List of grades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GradeResponse'

  /api/professors/grades/{id}:
    get:
      tags:
        - Professors
      summary: Get grade by ID
      operationId: getGradeById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Grade details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GradeResponse'
    put:
      tags:
        - Professors
      summary: Update grade by ID
      operationId: updateGrade
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGradeRequest'
      responses:
        '200':
          description: Updated grade
    delete:
      tags:
        - Professors
      summary: Delete grade by ID
      operationId: deleteGrade
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Grade deleted

components:
  schemas:
    StudentRequest:
      type: object
      properties:
        name:
          type: string
        catalogueCode:
          type: string
        studentCode:
          type: string
      required:
        - name
        - catalogueCode
    CreateStudentRequest:
      allOf:
        - $ref: '#/components/schemas/StudentRequest'
    UpdateStudentRequest:
      allOf:
        - $ref: '#/components/schemas/StudentRequest'
        - type: object
          properties:
            id:
              type: integer
    StudentResponse:
      allOf:
        - $ref: '#/components/schemas/StudentRequest'
        - type: object
          properties:
            id:
              type: integer

    DisciplineRequest:
      type: object
      properties:
        name:
          type: string
        disciplineCode:
          type: string
      required:
        - name
    CreateDisciplineRequest:
      allOf:
        - $ref: '#/components/schemas/DisciplineRequest'
    UpdateDisciplineRequest:
      allOf:
        - $ref: '#/components/schemas/DisciplineRequest'
        - type: object
          properties:
            id:
              type: integer
    DisciplineResponse:
      allOf:
        - $ref: '#/components/schemas/DisciplineRequest'
        - type: object
          properties:
            id:
              type: integer

    GradeRequest:
      type: object
      properties:
        semesterId:
          type: integer
        disciplineId:
          type: integer
        studentId:
          type: integer
        professorCode:
          type: string
        gradeValue:
          type: number
          format: float
        date:
          type: string
          format: date
      required:
        - semesterId
        - disciplineId
        - studentId
        - professorCode
        - gradeValue
    CreateGradeRequest:
      allOf:
        - $ref: '#/components/schemas/GradeRequest'
    UpdateGradeRequest:
      allOf:
        - $ref: '#/components/schemas/GradeRequest'
        - type: object
          properties:
            id:
              type: integer
    GradeResponse:
      allOf:
        - $ref: '#/components/schemas/GradeRequest'
        - type: object
          properties:
            id:
              type: integer

    ProfessorResponse:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
        disciplineCode:
          type: string
        professorCode:
          type: string

    GradesStudentResponse:
      type: object
      properties:
        semesterName:
          type: string
        disciplineCode:
          type: string
        professorCode:
          type: string
        gradeValue:
          type: number
          format: float
        date:
          type: string
          format: date